feat(cli): route migration/db commands through the control-api seam
Move the shared policy modules (plan-resolution, ref-advancement,
contract-space aggregate loader, seed phase, contract-at error mapping,
status overlay) and the migration-check/list/new/plan/ref/migrate-show
cores into src/control-api/operations/, add the client-free seam
operations (ref-resolution, graph-queries, invariants, caught-errors,
contract-snapshot-resolution, refs, migration-graph entries), and
reroute every command module so no file under src/commands/ imports
@internal/migration-tools at runtime. Export the new surface from
src/exports/control-api.ts and enforce the boundary with a
dependency-cruiser rule (tsPreCompilationDeps: "specify" labels pure
type-only edges).

Refs: TML-3173
