+++ packages/1-framework/3-tooling/cli/src/commands/migration-log.ts
@@ -1,11 +1,11 @@
 import { loadConfig } from '@internal/config-loader';
 import type { LedgerEntryRecord } from '@internal/contract/types';
-
 import { ifDefined } from '@internal/utils/defined';
 import { notOk, ok, type Result } from '@internal/utils/result';
 import { Command } from 'commander';
 import { createControlClient } from '../control-api/client';
-import { CliStructuredError, errorUnexpected, requireLiveDatabase } from '../utils/cli-errors';
+import { mapCaughtMigrationError } from '../control-api/operations/caught-errors';
+import { type CliStructuredError, errorUnexpected, requireLiveDatabase } from '../utils/cli-errors';
 import {
   addGlobalOptions,
   maskConnectionUrl,
@@ -87,7 +87,8 @@ export async function executeMigrationLogCommand(
     const ledger = await client.readLedger();
     return ok(ledger);
   } catch (error) {
-    if (CliStructuredError.is(error)) return notOk(error);
+    const mapped = mapCaughtMigrationError(error);
+    if (mapped) return notOk(mapped);
     return notOk(
       errorUnexpected(error instanceof Error ? error.message : String(error), {
         why: `Failed to read migration log: ${error instanceof Error ? error.message : String(error)}`,
