+++ packages/1-framework/3-tooling/cli/src/control-api/operations/caught-errors.ts
@@ -1,0 +1,13 @@
+/**
+ * Classifies errors caught around migration-tools calls so commands never import the MigrationToolsError class directly.
+ */
+
+import { CliStructuredError } from '../../utils/cli-errors';
+
+/** CliStructuredError (including MigrationToolsError) → identity; anything else → null (caller rethrows/wraps). */
+export function mapCaughtMigrationError(error: unknown): CliStructuredError | null {
+  if (CliStructuredError.is(error)) {
+    return error;
+  }
+  return null;
+}
