+++ packages/1-framework/3-tooling/cli/src/control-api/operations/contract-at-errors.ts
@@ -6,7 +6,7 @@ import {
   errorFileNotFound,
   errorSnapshotMissing,
   errorUnexpected,
-} from './cli-errors';
+} from '../../utils/cli-errors';
 
 export function mapContractAtError(
   error: unknown,
