+++ packages/1-framework/3-tooling/cli/src/control-api/operations/contract-space-aggregate-loader.ts
@@ -14,9 +14,9 @@ import { EMPTY_CONTRACT_HASH } from '@internal/migration-tools/constants';
 import { MigrationToolsError } from '@internal/migration-tools/errors';
 import { blindCast } from '@internal/utils/casts';
 import { notOk, ok, type Result } from '@internal/utils/result';
-import { CliStructuredError, errorUnexpected } from './cli-errors';
-import { readContractEnvelope, resolveContractPath } from './command-helpers';
-import { toDeclaredExtensionsFromRaw } from './extension-pack-inputs';
+import { CliStructuredError, errorUnexpected } from '../../utils/cli-errors';
+import { readContractEnvelope, resolveContractPath } from '../../utils/command-helpers';
+import { toDeclaredExtensionsFromRaw } from '../../utils/extension-pack-inputs';
 
 const CONTRACT_SPACES_DOCS_URL = 'https://pris.ly/contract-spaces';
