+++ packages/1-framework/3-tooling/cli/src/control-api/operations/ref-advancement.ts
@@ -1,8 +1,10 @@
 import { readFile } from 'node:fs/promises';
 import { writeContractSnapshot } from '@internal/migration-tools/contract-snapshot-store';
-import { errorInvalidRefName } from '@internal/migration-tools/errors';
+import { errorInvalidRefName, MigrationToolsError } from '@internal/migration-tools/errors';
 import { validateRefName, writeRef } from '@internal/migration-tools/refs';
 import { ifDefined } from '@internal/utils/defined';
+import { notOk, ok, type Result } from '@internal/utils/result';
+import type { CliStructuredError } from '../../utils/cli-errors';
 
 export interface ContractIR {
   readonly contract: unknown;
@@ -84,4 +86,77 @@ export async function buildRefAdvancementFields(options: {
     options.contractIR,
   );
   return { advancedRef, plannedAdvanceRef: null };
+}
+
+export interface ResolveRefAdvancementFieldsOptions {
+  readonly advanceRef?: string;
+  readonly db?: string;
+  readonly refsDir: string;
+  readonly migrationsDir: string;
+  readonly contractJson: Record<string, unknown>;
+  /** Path whose sibling .d.ts readContractIR derives (contract.json path). */
+  readonly contractJsonPath: string;
+  readonly mode: 'plan' | 'apply';
+  readonly hash: string;
+}
+
+/**
+ * Full ref-advancement phase for db init/update: ok({advancedRef:null, plannedAdvanceRef:null})
+ * when computeRefAdvancementName is null; else readContractIR + buildRefAdvancementFields with
+ * MigrationToolsError mapped, other errors rethrown.
+ */
+export async function resolveRefAdvancementFields(
+  options: ResolveRefAdvancementFieldsOptions,
+): Promise<Result<RefAdvancementFields, CliStructuredError>> {
+  if (
+    computeRefAdvancementName({
+      ...ifDefined('advanceRef', options.advanceRef),
+      ...ifDefined('db', options.db),
+    }) === null
+  ) {
+    return ok({ advancedRef: null, plannedAdvanceRef: null });
+  }
+  try {
+    const contractIR = await readContractIR(options.contractJson, options.contractJsonPath);
+    const fields = await buildRefAdvancementFields({
+      ...ifDefined('advanceRef', options.advanceRef),
+      ...ifDefined('db', options.db),
+      refsDir: options.refsDir,
+      migrationsDir: options.migrationsDir,
+      contractIR,
+      mode: options.mode,
+      hash: options.hash,
+    });
+    return ok(fields);
+  } catch (error) {
+    if (MigrationToolsError.is(error)) {
+      return notOk(error);
+    }
+    throw error;
+  }
+}
+
+/** migrate's --advance-ref tail: executeRefAdvancement with MigrationToolsError passed through, others rethrown. */
+export async function advanceRefSafely(args: {
+  readonly refsDir: string;
+  readonly migrationsDir: string;
+  readonly name: string;
+  readonly hash: string;
+  readonly contractIR: ContractIR;
+}): Promise<Result<{ readonly name: string; readonly hash: string }, CliStructuredError>> {
+  try {
+    const advanced = await executeRefAdvancement(
+      args.refsDir,
+      args.migrationsDir,
+      args.name,
+      args.hash,
+      args.contractIR,
+    );
+    return ok(advanced);
+  } catch (error) {
+    if (MigrationToolsError.is(error)) {
+      return notOk(error);
+    }
+    throw error;
+  }
 }
